7. Расстояние между точками в системе координат [#188 #258515]

+++ src/main/java/ru/job4j/condition/Point.java
@@ -13,7 +13,7 @@ public class Point {
         int y1 = 0;
         int x2 = 2;
         int y2 = 0;
-        int result = Point.distance(x1, y1, x2, y2);
+        double result = Point.distance(x1, y1, x2, y2);
 
         System.out.println("result (" + x1 + "," + y1 + ") to (" + x2 + "," + y2 + ") = " + result);
     }
